CueController - Increased max Cue size from uint16 to uint32.
CueController - Revised `read()` logic to check the buffered Cue size.

+++ src/cue/animationcuehandler.cpp
@@ -10,7 +10,7 @@ namespace PixelMaestro {
 
 	// Animation-specific Cues
 	uint8_t* AnimationCueHandler::set_fire_options(uint8_t section_num, uint8_t layer_num, uint8_t multiplier) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetFireOptions,
 			section_num,
@@ -22,7 +22,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_lightning_options(uint8_t section_num, uint8_t layer_num, uint8_t num_bolts, int8_t drift, uint8_t fork_chance) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetLightningOptions,
 			section_num,
@@ -36,7 +36,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_plasma_options(uint8_t section_num, uint8_t layer_num, float size, float resolution) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetPlasmaOptions,
 			section_num,
@@ -50,7 +50,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_radial_options(uint8_t section_num, uint8_t layer_num, uint8_t resolution) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetRadialOptions,
 			section_num,
@@ -62,7 +62,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_sparkle_options(uint8_t section_num, uint8_t layer_num, uint8_t threshold) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetSparkleOptions,
 			section_num,
@@ -74,7 +74,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_wave_options(uint8_t section_num, uint8_t layer_num, int8_t skew) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetWaveOptions,
 			section_num,
@@ -87,7 +87,7 @@ namespace PixelMaestro {
 
 	// General-purpose Cues
 	uint8_t* AnimationCueHandler::set_cycle_index(uint8_t section_num, uint8_t layer_num, uint8_t cycle_index) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetCycleIndex,
 			section_num,
@@ -99,7 +99,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_fade(uint8_t section_num, uint8_t layer_num, bool fade) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetFade,
 			section_num,
@@ -111,7 +111,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_orientation(uint8_t section_num, uint8_t layer_num, Animation::Orientation orientation) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetOrientation,
 			section_num,
@@ -123,7 +123,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_palette(uint8_t section_num, uint8_t layer_num, const Palette& palette) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetPalette,
 			section_num,
@@ -137,7 +137,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_reverse(uint8_t section_num, uint8_t layer_num, bool reverse) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetReverse,
 			section_num,
@@ -149,7 +149,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::set_timer(uint8_t section_num, uint8_t layer_num, uint16_t interval, uint16_t delay) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::SetTimer,
 			section_num,
@@ -163,7 +163,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::start(uint8_t section_num, uint8_t layer_num) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::Start,
 			section_num,
@@ -174,7 +174,7 @@ namespace PixelMaestro {
 	}
 
 	uint8_t* AnimationCueHandler::stop(uint8_t section_num, uint8_t layer_num) {
-		uint16_t index = start_cue(
+		uint32_t index = start_cue(
 			(uint8_t)CueController::Handler::AnimationCueHandler,
 			(uint8_t)Action::Stop,
 			section_num,
